# Conflicts:
#	src/main/resources/templates/checkout.html

+++ src/main/java/com/team1/efep/models/request_models/VNPayRequest.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Builder
 public class VNPayRequest {
-    private long amount;
+    private float amount;
 }
